web/django/requirements.txt
Django>=4.2.15
daphne
Twisted[tls,http2]==24.10.0
psycopg2-binary==2.9.6
hvac
django-redis
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
django-allauth==65.2.0
django-cleanup==9.0.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
phonenumbers==8.13.46
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
qrcode==7.4.2
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
shortuuid==1.0.13
sqlparse==0.5.1
tomli==2.0.2
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.1
